+++ app/models/point.rb
@@ -1,18 +1,20 @@
 # frozen_string_literal: true
 
 class Point < ApplicationRecord
-  reverse_geocoded_by :latitude, :longitude
+  include Nearable
+  include Distanceable
 
   belongs_to :import, optional: true, counter_cache: true
   belongs_to :visit, optional: true
   belongs_to :user
 
-  validates :latitude, :longitude, :timestamp, presence: true
-  validates :timestamp, uniqueness: {
-    scope: %i[latitude longitude user_id],
+  validates :timestamp, :lonlat, presence: true
+  validates :lonlat, uniqueness: {
+    scope: %i[timestamp user_id],
     message: 'already has a point at this location and time for this user',
     index: true
   }
+
   enum :battery_status, { unknown: 0, unplugged: 1, charging: 2, full: 3 }, suffix: true
   enum :trigger, {
     unknown: 0, background_event: 1, circular_region_event: 2, beacon_event: 3,
@@ -45,6 +47,14 @@ class Point < ApplicationRecord
 
   def reverse_geocoded?
     reverse_geocoded_at.present?
+  end
+
+  def lon
+    lonlat.x
+  end
+
+  def lat
+    lonlat.y
   end
 
   private