+++ spec/models/point_spec.rb
@@ -12,6 +12,7 @@ RSpec.describe Point, type: :model do
     it { is_expected.to validate_presence_of(:latitude) }
     it { is_expected.to validate_presence_of(:longitude) }
     it { is_expected.to validate_presence_of(:timestamp) }
+    it { is_expected.to validate_presence_of(:lonlat) }
   end
 
   describe 'scopes' do